refactor: updated css for home screen
